requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple
gunicorn
custard~=1.0.6
jsonpath~=0.82
pyjwt>=2.4.0
loguru~=0.6.0
requests~=2.27.0
fastapi~=0.78.0
uvicorn[standard]
supervisor~=4.2.4
pydantic~=1.9.0
starlette>=0.25.0
typing-extensions~=4.3.0
aiohttp~=3.8.1
awaits~=0.0.1